LayoutTests/http/tests/security/contentSecurityPolicy/script-src-self-blocked-03-expected.txt

CONSOLE MESSAGE: Refused to load the script 'https://127.0.0.1:8443/security/contentSecurityPolicy/resources/script.js' because it violates the following Content Security Policy directive: "script-src 'self'".



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
